requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cryptography==2.3.1
Django==2.0.6
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
djangorestframework==3.8.2
Fabric==1.13.2
Fabric3==1.13.1.post1
paramiko==2.4.2
psycopg2-binary
pyasn1==0.4.4
git+https://github.com/Moons15/django-rest-framework-docs
pycparser==2.19
PyNaCl==1.3.0
pytz==2018.7
six==1.11.0
gunicorn==19.6.0

